Add the option to add ratings

+++ backend/DTOs/RatingForObjectDTO.cs
@@ -4,7 +4,7 @@ namespace R8titAPI.Dtos
     {
         public string CategoryName { get; set; } = "";
         public bool Global { get; set; }
-        public int RatingValue { get; set; }
+        public decimal RatingValue { get; set; }
         public int RatingCategoryId { get; set; }
         public int UserId { get; set; }
         public int ImageId { get; set; }
